renamed sanitisevalue js file
