+++ src/app/garfile/departure/index.njk
@@ -202,7 +202,7 @@
 </div>
 
 <script type="text/javascript" src="/javascripts/accessible-autocomplete.min.js"></script>
-<script type="text/javascript" src="/javascripts/sanitise-value.js"></script>
+<script type="text/javascript" src="/javascripts/sanitisevalue.js"></script>
 <script type="text/javascript">
   accessibleAutocomplete.enhanceSelectElement({
     selectElement: document.querySelector('#departurePort'),
